+++ run-miri-tests.sh
@@ -3,9 +3,9 @@ set -e
 
 MIRI_NIGHTLY=nightly-$(curl -s https://rust-lang.github.io/rustup-components-history/x86_64-unknown-linux-gnu/miri)
 echo "Installing latest nightly with Miri: $MIRI_NIGHTLY"
-rustup default "$MIRI_NIGHTLY"
+rustup toolchain add "$MIRI_NIGHTLY"
 
-rustup component add miri
-cargo miri setup
+rustup component add miri --toolchain "$MIRI_NIGHTLY"
+rustup run "$MIRI_NIGHTLY" -- cargo miri setup
 
-cargo miri test 
+rustup run "$MIRI_NIGHTLY" -- cargo miri test