+++ src/imp_std.rs
@@ -5,11 +5,11 @@
 
 use std::{
     cell::UnsafeCell,
-    panic::{UnwindSafe, RefUnwindSafe},
     marker::PhantomData,
+    panic::{RefUnwindSafe, UnwindSafe},
     ptr,
+    sync::atomic::{AtomicBool, AtomicUsize, Ordering},
     thread::{self, Thread},
-    sync::atomic::{AtomicUsize, AtomicBool, Ordering},
 };
 
 #[derive(Debug)]
@@ -235,7 +235,10 @@ impl Drop for Finish<'_> {
 // These test are snatched from std as well.
 #[cfg(test)]
 mod tests {
-    use std::{panic, sync::mpsc::channel, thread};
+    use std::panic;
+    #[cfg(not(miri))] // miri doesn't support threads
+    use std::{sync::mpsc::channel, thread};
+
     use super::OnceCell;
 
     impl<T> OnceCell<T> {
@@ -256,6 +259,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(not(miri))] // miri doesn't support threads
     fn stampede_once() {
         static O: OnceCell<()> = OnceCell::new();
         static mut RUN: bool = false;
@@ -292,6 +296,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(not(miri))] // miri doesn't support panics
     fn poison_bad() {
         static O: OnceCell<()> = OnceCell::new();
 
@@ -313,6 +318,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(not(miri))] // miri doesn't support panics
     fn wait_for_force_to_finish() {
         static O: OnceCell<()> = OnceCell::new();